+++ CHANGELOG.md
@@ -6,6 +6,23 @@ All user visible changes to this project will be documented in this file. This p
 
 
 
+## [0.10.0] · 2026-??-??
+[0.10.0]: /../../tree/v0.10.0
+
+[Diff](/../../compare/v0.9.0...v0.10.0) | [Milestone](/../../milestone/66)
+
+### Fixed
+
+- UI:
+    - Media panel:
+        - Enabled video sometimes not being seen by participants. ([#1602])
+        - Call buttons dragged from dock to panel are being left hanging on screen. ([#1602])
+
+[#1602]: /../../pull/1602
+
+
+
+
 ## [0.9.0] · 2026-02-05
 [0.9.0]: /../../tree/v0.9.0